+++ tests/test_grid3d/test_grid_properties.py
@@ -186,6 +186,16 @@ def test_scan_dates():
     assert dl[2][1] == 20000201
 
 
+def test_dates_from_restart():
+    """A simpler static method to scan dates in a RESTART file"""
+    t1 = xtg.timer()
+    dl = GridProperties.scan_dates(RFILE1, datesonly=True)  # no need to make instance
+    t2 = xtg.timer(t1)
+    print("Dates scanned in {} seconds".format(t2))
+
+    assert dl[4] == 20000401
+
+
 def test_scan_keywords():
     """A static method to scan quickly keywords in a RESTART/INIT/*GRID file"""
     t1 = xtg.timer()